ajouter un equipement

+++ backend/src/main/java/com/itsupport/backend/repository/EquipementRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EquipementRepository extends JpaRepository<Equipement, Integer> {
+public interface EquipementRepository extends JpaRepository<Equipement, Long> {
 }
